refactor: share duplicated code between checkout payment components
